ON_DRIVER_RESTARTED Event infrastructure
This adds a new event, `DriverConfiguration.ON_DRIVER_RESTARTED` to REEF that will be called in lieu of `ON_DRIVER_STARTED` when a (failed) Driver is restarted by the resource manager. This event is *not* called yet and this PR is part of a series of changes to add support for Driver restarts.