Dependency upgrades to clean up SLF4J logs
Closes #929

SLF4J binds to a logging framework of choice by putting an adapter for that logging framework onto the classpath.  Turns out that we had several such adapters to Log4J on the class path, all of which are coming from the various hadoop subprojects.

This PR filters out these transitive dependencies. Also, it adds an adapter from SLF4J to the JDK logging library to the classpath of all leaf projects (examples and tests).

Note that adding them to inner projects of the dependency tree (e.g. `reef-common`) would be very bad, as we'd recreate the issue we had with inheriting Hadoop's choices for our consumers.