Updates to the event generation in the local runtime
Closes #986

Different than assumed in #986, the test failure of the `EvaluatorExitTest` was not because of the fact that the local runtime and REEF share the same thread, but because the local runtime did produce events in the wrong order: It first released the resources and then generated an event about that to the Driver. It needs to be the other way around to make sure the Driver receives the `FailedEvaluator` in time. This PR fixes that.