[REEF-80][HotFix] Log exceptions thrown as part of (de-)allocations
This PR changes `YarnContainerManager` such that it logs exceptions thrown as part of logging container (de-)allocations instead of re-throwing them as `RuntimeException`.

This is NOT a fix for the underlying issue, but merely a less invasive failure behavior: Where current master crashes, the code with this patch doesn't support driver restarts anymore in case of failure.