Today SAS uses Launch/JobClient in bridge project to generate driver config file, and submit it to REEF Launcher. This requires the clients to run Java bridge Launch offline with their configuration parameters. It is easy to cause out of sync for driver configuration settings.
This PR is to generate a skeleton of  driver configuration at Java side with Java bridge JobDriver configuration settings only.  This would leave the flexibility to the client so that they can deserialize the driver configuration, and set additional parameters and then serialize the result into driver config file again.  And we have provided an API for it.

To make the configuration modularized, this PR also separated HttpServer configuration and NameServer configuration from driver configuration file when creating the skeletons. They would become building blocks. Client can choose what they want in the driver and merge them as need. This would make Http Server and Name server become optional instead of build-in in a .Net/Java hybrid driver.

The driver skeleton configuration creation has been put as part of the build and the configuration files are packed into the jar file.

There are still some static and dynamic portion of driver configurations set at Reef Runtime. We need to abstract them out and add them into the config file so that clients can call REST API directly with the complete driver config. This work is not covered in this PL.

This is to resolve issue #54. https://issues.apache.org/jira/i#browse/REEF-54

Author: Julia Wang
email: jwang98052@yahoo.com