[REEF-180] Automate pushing REEF.NET NuGet packages to Nuget.org
Added PushPackages and NuGetRepository properties to build.props. When
PushPackages is set to true and Configuration is set to Release, the
NuGet packages that are generated will be pushed to NuGet.Org.  You can
also change the repository that the packages are pushed to using the
NuGetRepository property.

If pushing to NuGet.org, you must set the NuGetApiKey environment
variable to be a valid API key.  You must also be an owner of the
REEF.NET packages on NuGet.org.