[REEF-205] Add Org.Apache.REEF.Bridge.JAR
This adds a new project to the .NET solution:
`Org.Apache.REEF.Bridge.JAR`. Upon build, it checks whether the shaded
jar for `reef-bridge-java` exists. If not, it will be built. The build
will fail if M2_HOME isn't set and copying the JAR to the output folder.

Also, this PR adds the shaded jar to `reef-bridge-java`.

JIRA:
  [REEF-205](https://issues.apache.org/jira/browse/REEF-205)

Pull Request:
  This closes #114

Author:    Markus Weimer <weimer@apache.org>