[REEF-232]  Resolving race condition in Communication Group
When CommunicationGroupNetworkObserver receives message from other nodes, handlers for operators may not be registered yet. This race condition may cause exception therefore resulting in sockets to be closed. When some nodes are not able to receive message from children or parents, entire communication get stuck. This can be repro when adding more nodes with tree topology.

This PR is to resolve this issue by adding waiting and retry for handlers to be registered before processing the message.

JIRA: REEF-232. (https://issues.apache.org/jira/browse/REEF-232)

This closes #

Author: Julia Wang  Email: jwang98052@yahoo.com