[REEF-174]  Use type instead of instance for Codec in MpiDriver
Currently, a Codec instance is passed to the specifications in
MpiDriver. What we really need is the type that will be bound to Tang
Configuration. This change is to pass generic type into MdiDriver and
Specification instead so that any type of ICodec can be passed in. At
the meantime, we limit the type to be an instance of ICodec to ensure
correct type is passed. If not, catching the error at build time.
IOperatorSpec, Itopology, ICommunicationGroupDriver API and
implementations are modified accordingly.

Test cases are updated as well.

JIRA:
  [REEF-174](https://issues.apache.org/jira/browse/REEF-174)

Pull Request:
  This closes #132