[REEF-246] Add LocalAddressProvider
This PR adds the concept of a `LocalAddressProvider` that provides the
address which network services shall bind to. This new interface
replaces `NetUtils.getLocalAddress()`.

This adds two implementations for this new interface:

  * `LegacyLocalAddressProvider` which uses the code found prior in
    `NetUtils.getLocalAddress()`.
  * `HostnameBasedLocalAddressProvider`, which uses the same code Apache
    Hadoop MapReduce uses for this.

The default implementation for `LocalAddressProvider` is
`HostnameBasedLocalAddressProvider`. `NetUtils.getLocalAddress()` has
been deprecated and uses the default implementation internally.

The remainder of the changes are dealing with the fallout of this
change. In fact, this is the majority of the changes. Most notably,
this removes the `bindVolatile` on the clock injector in `LaunchClass`.
This means that we now potentially instantiate two `RemoteManager`s.
[REEF-269] has been filled to track the cleanup.

JIRA:
  [REEF-246](https://issues.apache.org/jira/browse/REEF-246)

Pull Request:
  This closes #155