[REEF-275] Only use ConfigurationProvider for JVM Evaluators
In `AllocatedEvaluatorImpl#launch`, [REEF-268] added a new call to
`AllocatedEvaluatorImpl#makeRootContextConfiguration`. That method
merges the given `Configuration` with the ones provided by the
`ConfigurationProvider` instances configured.

In the case of a CLR Evaluator launch, this means that we attempt to
merge JVM Configuration with CLR Configuration. Which is not (yet)
supported in Tang.

REEF-273 and REEF-274 will need to add a similar mechanism in .NET to
the one REEF-268 introduced for Java. Hence, this PR adds a check of the
type of Evaluator launched and only merge the configurations when we
indeed have a JVM launch at hand.

Also, this PR removes a few `@Parameter` annotations on a constructor of
`NameServerImpl` that isn't actually injectable.

JIRA:
  [REEF-275](https://issues.apache.org/jira/browse/REEF-275)

Pull Request:
  This closes #158

Author:    Markus Weimer <weimer@apache.org>