[REEF-344] Cleaned up LibLoader
After this change, LibLoader will only load the Bridge DLL from the
`reef/local` or `reef/global` folder. Preference is given to
`reef/local`.

If the DLL is found but can't be loaded, the bridge will now crash. This
is to avoid situations in which users believe they load
the version in `reef/local` while the system in fact silently switches
to `reef/global`.

Loading the DLL from the REEF JAR is no longer supported after this
change.

JIRA:
  [REEF-344](https://issues.apache.org/jira/browse/REEF-344)

Pull Request:
  This closes #204

Author:    Markus Weimer <weimer@apache.org>