[REEF-368]: Avoid holding Wake schedule lock while calling event handlers
This removes the call to handle an idle event to before the schedule
lock is taken.

This should avoid the deadlock situation seen in REEF-294, where a
separate thread holds a lock needed by the idle event handler, and then
waits for the schedule lock. Now the idle event handler call will not be
holding the schedule lock.

JIRA:
  [REEF-368] https://issues.apache.org/jira/browse/REEF-368

Pull Request:
  This closes #218