[REEF-390] Pass Service Configuration to Evaluator as a string (.NET)
Currently, some of the Nodes were dropped by proto buffer libary when
Java deserializes a class hierarchy created in C#, resulting in Java
driver not be able to deserialize service configurations passed from C#
which use those classes. A typical example is nested generic types for
ICodec which client would like to pass to the evaluator as part of the
service config.

This PR:
  * wrap the entire service config as a serialized string in service
    config when serializing it cross the bridge
  * Updated ServiceConfiguration to inject the named parameter to get
    the string config and deserialze it
  * Test cases are added for the change

JIRA:
  [REEF-390](https://issues.apache.org/jira/browse/REEF-390)

Pull Request:
  This closes #228