REEF-409: Pass the rackName from the runtime to the top of REEF stack
This work bubbles up the rack information from the runtime (after the resource is allocated) towards the evaluators allocation handlers.
The ResourceAllocationEvent interface now includes an optional rack name, which is populated by the respective runtimes when the resource is allocated.
The ProcessContainer is created with a null rack for now, i.e. the default is being used.
Added a unit test that tests the whole cycle, checking if the default rack is being returned.

JIRA:
  [REEF-409] (https://issues.apache.org/jira/browse/REEF-409)

This closes