[REEF-427] Add IMRU API and single threaded implementation
This adds `Org.Apache.REEF.IMRU.API` which contains the API for IMRU as well as
`Org.Apache.REEF.IMRU.InProcess` which is a single-threaded implementation for
testing.

Also, this change adds a very simple example IMRU program that counts the number
of mappers in the job. This example is used for the first test in
`Org.Apache.REEF.IMRU.Tests`

JIRA:
  [REEF-427](https://issues.apache.org/jira/browse/REEF-427)

Pull Request:
  This closes #