[REEF-453]  Race condition in GroupCommunication Client initialization
In GroupCommClient constructor, we register the task node to Name Server first, then instantiate CommunicationGroupClient, which will instantiate Operators and register handlers. In Operator constructors, we check if its parents and children are all registered based on the topology. There is a chance that a parent or a child was registered, but haven't completed the initialization, resulting a race condition.

This PR moves the registration in GroupCommClient as the last line in the constructor. And movs the initialize checking out of the Operator constructors. Initialize will be called when client gets Operator from CommunicationGroupClient. This would ensure all the registration happens after constructors. And registration checking happens after all the constructors.

JIRA: REEF-453(https://issues.apache.org/jira/browse/REEF-453)

This closes #

Author: Julia Wang  Email: jwang98052@yahoo.com