[REEF-454] Allow for per-job queues
This adds `YarnDriverConfiguration` that can be filled out by an application and
merged into their DriverConfiguration to set the queue to submit the job to.

This also makes sure the `YarnClientConfiguration.YARN_QUEUE_NAME` is actually
used (it wasn't). It now provides the default for when the new per-job setting
isn't used.

Lastly, this change deprecates `JobSubmissionEvent.getQueue()` which was never
used in the first place. Its purpose is now served by the mechanism introduced
here.

JIRA:
  [REEF-454](https://issues.apache.org/jira/browse/REEF-454)

Pull Request:
  This closes #279

Author:    Markus Weimer <weimer@apache.org>