[REEF-466]  Friend access was granted with public key but the module is not signed
There is an error in debug build as friend access was granted with public key but the module is not signed.
This PR is to sign debug build as well to avoid the issue so that both debug and release build can share the same AssemblyInfo file.

JIRA: REEF-466(https://issues.apache.org/jira/browse/REEF-466)

This closes #

Author: Julia Wang  Email: jwang98052@yahoo.com