[REEF-468] Support multiple compute and data requests in DataLoader
This commit allows to request a user-defined number of compute and data
evaluators to the DataLoader. Each evaluator can be requested to run in
particular locations, using the rack information. As there are some
clients of the DataLoadingRequestBuilder, we maintain its previous
functionality, though we marked several fields as deprecated. There is
still pending to implement the different strategies on how to assign
partitions to the different evaluators. For now, we use the previous
implementation (tries to match the host, if not, assigns them randomly).
Needed to update EvaluatorRequest.Builder to make "cloning" easier.

JIRA:
  [REEF-468](https://issues.apache.org/jira/browse/REEF-468)

Pull Request:
  This closes #287