[REEF-469] Assign data splits to specific evaluators
This commit allows to assign data splits to the different evaluators using the DataLoadingService.
Several changes were done. The main one is to inject a strategy into InputFormatLoadingService, which is in charge
of assigning splits to evaluators.
New data types needed to be created. For example, DataPartition, which is just a folder, but has a location field, where the user can specify where
she would like to load that data into (in certain node or rack).
We have LocationAwareJobConfs now, which binds the typical hadoop JobConf and the DataPartition objects.
Created external constructor for them.
Removing the EvaluatorToPartitionMapper class, most of its logic has been placed in the AbstractEvaluatorToSplitStrategy one.

JIRA:
  [REEF-469](https://issues.apache.org/jira/browse/REEF-469)

Pull Request:
  This closes