[REEF-509]  Remove Java ExecutorServiceConstructor from EvaluatorConfiguration for .Net Evaluator
Currently the Evaluator configuration contains ExecutorServiceConstructor which is not used in .Net evaluator and .Net Tang will be not able to understand this class. This PR created separate code path for EvaluatorConfiguration in java based on the Evaluator Process.

JIRA: REEF-509(https://issues.apache.org/jira/browse/REEF-509)

This closes #

Author: Julia Wang  Email: juliaw@apache.org