[REEF-574]  Eliminate Task Configuration deserialization/serialization at Java side
Currently Task configuration is passed over the bridge as a string. At Java side, it is deserialized into a Configuration object with C# class hierarchy. Before it is passed over to Evaluator, it is serialized again into a string.
This PR added overloading methods to allow the task configuration passing around as a string without deserialization. And then set it directly to evaluator configuration.
Those methods are added to the impl classes but not interface to avaoid being exposed as public interface.

JIRA: REEF-574(https://issues.apache.org/jira/browse/REEF-574)

This closes #

Author: Julia Wang  Email: juliaw@apache.org