[REEF-568] Work around the federated YARN node reports problem
This change is a work around for a federated YARN problem. Once the
corresponding JIRA is fixed, it should be reverted.
The issue is the following:
Just after initializing our yarn client library, we ask for the RUNNING
nodes in the cluster to populate our own Resource Catalog.
YARN replies with the nodes that belong to a 'random' sub-cluster; sometimes
with the nodes in the correct sub-cluster (where the containers will be
placed), but sometimes with other ones.
That causes the application to randomly fail.

This hot fix adds the information of the allocated containers to the
resource catalog if they are not there, when using using a federated YARN
environment. This last check is done by the "YarnFederation" configuration parameter.

JIRA:
  [REEF-568](https://issues.apache.org/jira/browse/REEF-568)

Pull Request:
  This closes