[REEF-589] REEF crashes when new nodes are added to the clusters dynamically
If an admin happens to dynamically add a new node to the cluster, and future allocations are done in that node, REEF crashes as it is not able to find that node in the catalog.
Though we found this problem using YARN with Federation, it is a problem that happens on all RMs but the Local Runtime, the latter doesn't support dynamic nodes additions.

The change involves adding a node to our catalog if an allocation event is received and the node is not already in the catalog.

JIRA:
  [REEF-589](https://issues.apache.org/jira/browse/REEF-589)

Pull Request:
  This closes