[REEF-589] Support node addition while REEF applications run
If an admin happens to dynamically add a new node to the cluster, and
future allocations are done in that node, REEF crashes as it is not able
to find that node in the catalog.

The change involves adding a node to our catalog if an allocation event
is received and the node is not already in the catalog.

JIRA:
  [REEF-589](https://issues.apache.org/jira/browse/REEF-589)

Pull Request:
  This closes #370