[REEF-613] DataLoadingRequestBuilder should use the desired number of input splits defined by the user
We are overriding the numberOfDesiredInputSplits defined by the user to 0,
when they invoke the DataLoadingRequestBuilder.setInputPath method.
This change allows to use the DataLoadingRequestBuilder.setInputPath method
and DataLoadingRequestBuilder.setNumberOfDesiredSplits safely.
We now set the correct number of desired splits to the DistributedDataSet object we create.
This also un-deprecates some of the API to avoid the overhead of
creating the multi data center strategy, and throws an exception
if setInputPath and setDistributedDataSet methods are both called (in any order).

JIRA:
  [REEF-613](https://issues.apache.org/jira/browse/REEF-613)

Pull Request:
  Closes #391