[REEF-620]  Adding language property in Avro Configuration
When the code in a language receives a configuration from another, it needs to know which language the configuration was created so that to determine which alias to choose.

This PR added a language property in Avro Configuration to carry this information. It also modified the places where the AvroConfiguration object is created in both Java and C#.

The evaluator.conf used in testing is a newly generated one from test runtime which contains language information. Test is updated to read language property from it after deserialization.

JIRA: REEF-620(https://issues.apache.org/jira/browse/REEF-620)

This closes #

Author: Julia Wang  Email: juliaw@apache.org