[REEF-518] Smarter VortexWorker-Tasklet scheduling
This addressed the issue by
  * Creating pluggable SchedulingPolicy for choosing a worker to launch a tasklet onto.
  * Implementing two types of SchedulingPolicy: FirstFit and Random.
  * Creating unit tests for testing these two policies.
  * Making changes in related classes to configure and use the SchedulingPolicy.

Zhengping Qian implemented the core logic. John Yang wrote tests, code for configurations, and JavaDocs.
Note that we changed the name of the class PendingTaskletScheduler to PendingTaskletLauncher to
make it clearer that the worker-tasklet matching is the responsibility of RunningWorkers.
In the code you can see that SchedulingPolicy is injected into RunningWorkers and used there.

We expect that later the PendingTaskletLauncher will be responsibile for addressing REEF-500.

JIRA:
  [REEF-518](https://issues.apache.org/jira/browse/REEF-518)