[REEF-651] Remove unnecessary logging information from reef-io.
Currently, many REEF-IO methods are logging unnecessary information when
entering and exiting. That seriously affects performance and can easily
cause OutOfMemory exceptions, e.g. when logging a big ML model. This
change cleans that up. Whenever entering or exiting is called, we won't
log any potential big object. If users decide to log that information in
their own application, so be it. But this shouldn't be done in the
communications library.

JIRA:
  [REEF-651](https://issues.apache.org/jira/browse/REEF-651)

Pull Request:
  Closes #420