[REEF-674] Made IEvaluatorDescriptor and its implementation immutable.
`EvaluatorRequestor` was changed to call the new constructor of
`EvaluatorDescriptorImpl`

This also removes the `FromString` method from both the interface and
the implementation. Its code is folded into the constructor which was
the only place it was called from.

Lastly, standard ReSharper cleanup has been applied to
`IEvaluatorDescriptor` and `EvaluatorDescriptorImpl`.

JIRA:
  [REEF-674](https://issues.apache.org/jira/browse/REEF-674)

Pull Request:
  This closes #469