[REEF-765] Add validation and logging on the Java side of the C# client
For YARN, this introduces the new class `YarnSubmissionFromCS`. It is used by
`YarnJobSubmissionClient` to parse, log and validate the command line parameters
passed from C#.

For local, this introduces the new class `LocalSubmissionFromCS`. It is used by
`LocalClient` to parse, log and validate the command line parameters passed from
C#.

JIRA:
  [REEF-765](https://issues.apache.org/jira/browse/REEF-765)

Pull Request:
  This closes #506

Author:    Markus Weimer <weimer@apache.org>