[REEF-831] Make YarnConfigurationUrlProviderTests stable
The issue was caused by a race between FileSystemWatcher holding a handle
to the config file being tested and the test itself trying to delete the
configuration file. Ideal solution is not to use physical files for these
tests, but that requires non-trivial amount of effort.

This addressed the issue by
 * Removing the config reload logic as it is not needed yet. We can
   revisit if this is needed by users.

JIRA:
  [REEF-831](https://issues.apache.org/jira/browse/REEF-831)

Pull request:
  This closes #565