[REEF-848] Update IPartition<IEnumerable<T>> to IPartition<T>
This change switches from IEnumerable<T> to T for the generic in IPartition.
Method GetPartitionHandle() also returns T.  It is up to the client to pass something like IEnumerable<U> for T.
Similarly, the return type from Deserialize() in FileDeSerializer also changed to T.
Test cases are updated as well.

JIRA:
  [REEF-848](https://issues.apache.org/jira/browse/REEF-848)

Pull request:
  This closes #572