[REEF-873] fix FileSystemPartitionInputDataSet id issue
This PR is to change the way to generate id for FileSystemPartitionInputDataSet. Original way was trying to derive the id from the input file name so that to make it easy to recognize. Given the fact that the input file name itself is can be random number, and the long file name may cause errors in parsing it, we would like to choose a simpler way to generate the id.

JIRA: [REEF-873](https://issues.apache.org/jira/browse/REEF-873)

This closes #