[REEF-1565] Make Context accessible from task
For fault tolerant, states should be saved at context level so that when the task is resubmitted, update task host would be able to restore the state from the context.

Currently, we start task from context. But inside task we are not able to access context. As ContextRuntime could have multiple instances, it is not created by injection.

This Jira is to make Task able to inject context where the task is running on. This is the first step for the state management.

JIRA: [REEF-1565](https://issues.apache.org/jira/browse/REEF-1565)
This closes