[REEF-1570] In-line AvroUtils and remove Avro dependency from O.A.R.Utilities
This moves AvroUtils from O.A.R.Utilities and in-lines its usage into O.A.R.Naming.Utilities
and O.A.R.Examples.MachineLearning.KMeans.codecs. Also, the Avro dependency from
O.A.R.Utilities is removed while an Avro dependency has been added to O.A.R.Examples.
The current usages have been tagged with a TemporaryFixAttribute so that we can fix it
appropriately when we are ready to do so.

JIRA:
    [REEF-1570](https://issues.apache.org/jira/browse/REEF-1570)

Pull Request:
    Closes #1127