[REEF-1641] Remove batch id and change the way to request maser/slave evalautors
Currently in IMRU, we have different spec for master vs mapper elevators. We use EvaluatorBatchId to distinguish which one is what we requested when receiving IAllocatedEvalauator. However, EvaluatorBatchId may not be supported in some environment such as HDI. So we need a way to decide which evaluator is for master.

The change in this PR is to request the first evaluator for master. After receiving it, record that one as master, then request the mappers.

An assumption is if master fails at any stage, fail the system.

* IMRU Driver and Evaluator Manager are updated
* Existing test cases are updated
* A new test case is added to simulate master evaluator failure during WaitingForEvaluator phase.

JIRA: [REEF-1641](https://issues.apache.org/jira/browse/REEF-1641)

This closes  #