[REEF-1641] Remove batch id and change the way to request maser/slave evalautors
Currently in IMRU, we have different spec for master vs mapper elevators.
We use EvaluatorBatchId to distinguish which one we requested when receiving IAllocatedEvaluator.
However, EvaluatorBatchId may not be supported in some environments such as HDInsight.
So we need a more reliable way to decide which evaluator is for master.

The change in this PR is to request the first evaluator for master.
After receiving it, record it as master, then request the mappers.

An assumption is that if master fails at any stage, we fail the system.

* IMRU Driver and Evaluator Manager are updated
* Existing test cases are updated
* A new test case is added to simulate master evaluator failure during WaitingForEvaluators phase.

JIRA:
  [REEF-1641](https://issues.apache.org/jira/browse/REEF-1641)

Pull request:
  This closes #1156