[REEF-1482] Driver does not exit even if all the task exit normally
Currently, when all the tasks and all the evaluators are completed,
sometimes driver still doen't shut down and hungs there forever.
This happens intermittently. When there are many nodes like 500 nodes
in IMRU runs in Yarn cluster, the issue can happen in every 2 or 3 runs.

The investigation shows there is a potential dead lock in ResourceManagerStatus.
This PR is to resolve this issue by reducing the scope of code under locks.

With the fixes, I have tested 10 times with 500 nodes in cluster, there is no repro any more.

JIRA:
  [REEF-1482](https://issues.apache.org/jira/browse/REEF-1482)

Pull request:
  This closes #1162