[REEF-1669]Removing YarnClient from YarnContainerManager
Currently we have both YarnClient and resourceManager (AMRMClientAsync) in YarnContainerManager. We call init() and start() for both of them. But inside driver, we are not able to get correct YarnConfiguration for YarnClient, that makes yarnClient.getNodeReports() fail to connect to RM.

In YarnContainerManager, we already have onNodesUpdated() which gives the list of the NodeReports on the fly.  yarnClient.getNodeReports() was called at beginning and it only gives a static list. So looks like we don't need to call it anyway.

Based on RM folks, REEF should not use YarnClient at driver side. This PR is to remove YarnClient from YarnContainerManager. It also added some logs for environment variables with a lower log level or debug.

JIRA: [REEF-1669](https://issues.apache.org/jira/browse/REEF-1669)
This closes  #