[REEF-1680] Increasing default retry time in WaitingForRegistration
In IMRU, update task doesn't need to do data load but mapper do.
That means we need to set sufficient retry time in WaitingForRegistration,
at least more than data loading time. Current setting is about 4 min.
But for big data, it can be 25 minutes.

Previously we hesitated to increase this number because WaitingForRegistration
happened in Task constructor. Before we get IRunningTask event, there is
no way to cancel a long waiting task in failure cases.

Now we have moved WaitingForRegistration out of task constructor and
added cancellation token so that we can cancel the waiting if failure happens.

This change increases the waiting time to 30min.

JIRA:
  [REEF-1680](https://issues.apache.org/jira/browse/REEF-1680)

Pull request:
  This closes #1194