[REEF-1680] Increasing default retry count in WaitingForRegistration
In IMRU, update task doesn't need to do data load but mapper do. That means we need to set sufficient retry times in WaitingForRegistratio, at least more than data loading time. Current setting is about 4 min. But for big data, it can be 25 minutes.
Previously we were hesitate to increase this number because WaitingForRegistration was in Task constructor. Before we get running task, there is no way to cancel a long waiting task in failure cases.
Now we have moved WaitingForRegistrationout of task constructor and added cancellation token. We are able to cancel the waiting if failure happens.
This PR is to increase the waiting time to 30min.

JIRA: [REEF-1680](https://issues.apache.org/jira/browse/REEF-1680)
This closes  #