[REEF-1682] Update TCP connection config values for IMRU example and test
Recent stress testing for IMRU FT shows when there are hundreds of nodes
with many failed and re-requested evaluators, the default TCP connection
retry time is not long enough for evaluators to connect to driver, as
driver might be busy handling a long event queue and each event handler
is locked in IMRU driver.

We need to set proper configuration values for IMRU example
which is used for running stress testing.

In some of the IMRU tests this configuration is set via task function config,
which has no effect. It should be set through configuration provider to work.

This change sets TCP connection configuration values for IMRU example
and fixes this setting in functional tests.

JIRA:
  [REEF-1682](https://issues.apache.org/jira/browse/REEF-1682)

Pull request:
  This closes #1197