[REEF-1685] Complete the Job properly if update/master task is completed from running state
In stress testing, we have seen the following scenario: master/update task is successfully completed,
and most of mapper tasks are also completed, but then driver receives an IFailedEvaluator event.
This would result in system transitioning to ShuttingDown state and doing an unnecessary retry.

When the driver receives ICompletedTask from master task which was in running state,
that means we have completed the calculation and result has been written to the output.
The driver should execute DoneAction to dispose of all the contexts and shut down the system.
After that, any FailedEvalutor/FailedTask events received should be ignored.

JIRA:
  [REEF-1685](https://issues.apache.org/jira/browse/REEF-1685)

Pull request:
  This closes #1200