[REEF-1687] Throw exception if task dispose throws exception after Call()
Currently in TaskRuntime, when Dispose a task in Close method, if exception happens, we throw exception. While when dispose the task after calling Call() method, we just log but swallow the exception. This is inconsistent. It also result in TestFailMapperTasksOnDispose randomly fails because it really depends on which code path is quicker that determines where the task Dispose is called.

When Task Dispose throw exception, that means some resources are not released properly. For fault tolerant, even if the task is returned from the Call(), either really completed the task or clsoedByDriver, if we are to reuse the evaluator to resubmit a task on it, it is not right. So the correct way is to throw exception to make evaluator fail, same as what we did in close() method.

This PR update both TaskRuntime and TestFailMapperTasksOnDispose

JIRA: [REEF-1687](https://issues.apache.org/jira/browse/REEF-1687)
This closes  #1203